Admin Product Management

+++ EcommerceBackend/src/main/java/com/learning/DTO/ProductRequest.java
@@ -4,14 +4,16 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductRequest {
     private String title;
     private String description;
-    private double price;
-    private double discountPrice;
+    private BigDecimal price;
+    private BigDecimal discountPrice;
     private double discountPercentage;
     private Integer stock;
     private String brand;
